Split field steps
