+++ frictionless/steps/field/field_remove.py
@@ -1,0 +1,35 @@
+from ...step import Step
+
+
+# NOTE:
+# Some of the following step can support WHERE/PREDICAT arguments (see petl)
+# Some of the following step use **options - we need to review/fix it
+
+
+class field_remove(Step):
+    """Remove field"""
+
+    code = "field-remove"
+
+    def __init__(self, descriptor=None, *, names=None):
+        self.setinitial("names", names)
+        super().__init__(descriptor)
+
+    # Transform
+
+    def transform_resource(self, resource):
+        table = resource.to_petl()
+        names = self.get("names")
+        for name in names:
+            resource.schema.remove_field(name)
+        resource.data = table.cutout(*names)
+
+    # Metadata
+
+    metadata_profile = {  # type: ignore
+        "type": "object",
+        "required": ["names"],
+        "properties": {
+            "names": {"type": "array"},
+        },
+    }
